ingest fractional shares

+++ ingest/src/protobuf/quotation.rs
@@ -11,16 +11,16 @@ pub struct QuotationMessageProto {
     pub bpu: i32,
     #[prost(int32, tag = "5")]
     pub bpf: i32,
-    #[prost(int32, tag = "6")]
-    pub bs: i32,
+    #[prost(double, tag = "6")]
+    pub bs: f64,
     #[prost(string, tag = "7")]
     pub ax: ::prost::alloc::string::String,
     #[prost(int32, tag = "8")]
     pub apu: i32,
     #[prost(int32, tag = "9")]
     pub apf: i32,
-    #[prost(int32, tag = "10")]
-    pub r#as: i32,
+    #[prost(double, tag = "10")]
+    pub r#as: f64,
     #[prost(message, optional, tag = "11")]
     pub ti: ::core::option::Option<::prost_types::Timestamp>,
     #[prost(string, repeated, tag = "12")]
